Update URI api
